get text highlight
