+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version 0.39.0 - Unreleased
+- Added setter for `ITextFrame.LeftMargin`, `ITextFrame.RightMargin`, `ITextFrame.TopMargin` and `ITextFrame.BottomMargin` properties [#385](https://github.com/ShapeCrawler/ShapeCrawler/issues/385)
+
 ## Version 0.38.0 - 2022-11-28
 - Added setter for `ITextFrame.AutofitType` property [#360](https://github.com/ShapeCrawler/ShapeCrawler/issues/360)
